PAL: Handle SSR notification delay case
If SSR notification is delayed in GSL, then GSL will return
error code for genuine failure, but the failure has happened
due to SSR, and hence, in PAL this failure will not be handled
as SSR, which will lead to incorrect cached stream state, while
handling SSR down cycle.

Hence, we need to check if SSR notification is landed in PAL,
then we can consider it as SSR case, and cache proper stream state.

Change-Id: I780240cda9c393ee6c026439734cf0ead61af2fe