pal: checkHapticsConcurrency in switchDevice for all HS usecases.
checkHapticsConcurrency for all HS usecases, because if LL stream
is active with haptics on speaker and HS, and then HS is removed and
music at 44.1 KHz resumed in between, and HS is inserted back again,
then device was configuring at 44.1 KHz, which is not correct. HS should
be updated to 48KHz because haptics is active. Fix this by checking
haptics concurrency even when there are streams present which share BE.

Change-Id: I465c80a8eb0a1eae25fe1573f387f55ed3fb2904