pal: Dynamically load ACD sound models
Without altering the PAL code, any new sound model
that needs to be tested/added can be easily implemented
by adding it to the resource manager xml file.

A few unnecessary or redundant functions are eliminated.

The list of models that must be loaded/unloaded is now
managed by unordered map.

Change-Id: I955860a1b050938d8f738a75d4ccb2e05ea237ac