pal: do forceDeviceSwitch for only suspended streams
During a2dpSuspend/a2dpCaptureSuspend, we mute the streams and route
to speaker/handset/handsetMic. We check all the active streams on inDev
during forceDeviceSwitch. This includes even the new stream which is created
after we process mute. This way we will be routing a stream which is not in
muted state, which will lead to audio leak on speaker.

This change ensures the routing of only the streams which were initially
active during a2dpSuspend/a2dpCaptureSuspend.

Change-Id: If592d634d990179a7366984d99860a7dc6302fe9