pal: refine readoffset usage in ring buffer and interface
 - update ring buffer read offset and unread size for all
related readers when detection is received.

 - start/end indices are updated to be offsets from
beginning of the keyword in the buffer.

 - ftrt size for a given detection is added as the offset
from begining of the ring buffer.

 - read offset stored in interface side indicates keyword
length including pre-roll. This will be used to advance
the read offset of the reader in the ring buffer when
client starts reading, depending on client has requested
for keyword or not.

Change-Id: Icc8124f4704819ddc3dcafcf711badfec10accd4