PAL: Add buffering state check while updating engine config on stop
In case of dual VA, if stream instance 1 is in buffering state, and stream instance 2
get unloaded, then while stopping the engine, we update the engine config, without
considering buffering stream as active stream, as reset the capture_requested_ flag,
due to which, for upcoming detections of stream 1, GSL engine will not go to buffering
state, and it will not read any data, which will lead to infinite loop in capi engine.

Now, adding buffering state check, to avoid mentioned scenario.

Change-Id: Ie6fc86039c4fa1a335f16de8becc4cc9865abd9d