Pal: Only set ec device to null when disabling same ec device
If stream was associated with one Rx device for EC ref, and later it got
associated with another Rx device for EC ref, then stream get request to
disable first Rx device, in this case, this disable request was ignored
from GSL side, but at stream side we set EC Rx device to null, due to
which later if stream get any new Rx device for EC ref, it was treated as,
stream is setting Ec for first time, and it leads the ref count to go out
of sync.

Now, added a check, to ensure either device was NULL, or the latest RX device
for which EC was set.

Change-Id: I1ec816b8967e0e021f217966bf48ca20bcbaff6c