Pal: Avoid switching UPD streams to Bluetooth devices.
During the forced device switch, all active streams
are switched to the new Bluetooth device connected.
UPD is one stream that needs to be excluded.
With this change, the stream will now be checked to
see if it is UPD and legitimate before being switched
to the new incoming device.
A generic function is created to do this check in
various stream device switch functions.

Change-Id: I8087ac96e7880c730e478530ef8d758fa32bf935