PAL: Update ec_ref_mutex to be recursive mutex
When external EC is connected and internal EC tries to get connected,
then it will not get connected, but cached, and when external EC got
disabled, then cached internal EC will try to connect, and all this
will happen in same thread, due to which, we get deadlock for gsl
engine's ec_ref_mutex.

Hence, updating the type of mutex, to recursive mutex, so that same
thread will not wait for same mutex, which it has locked earlier.

Change-Id: I5e649c3e5d391fbc5ee21a84bc4d0a3b092e2f53