pal: Check for Start Recognition during Device switch.
In case of dual VA, if the first stream is in buffering state,
and second stream is in loaded state, while handling connect
device event for first stream, it will try to start
recognition, but will fail as device is not actually
connected. It will get connected when connect event is called
for last stream but last stream will remain in loaded state,
hence detection for first stream will never get started.

Now, adding check if any stream was in active state,
then we start engine, while handling connection for last stream.

Change-Id: I0c6758e7e1942fc65b7236e9ef34a4e1babb603a