pal: set mDevices as BT A2DP in case of a2dpResume() during SSR DOWN
In case of SSR down event if a2dpSuspended = false sets, since sound
card state is offline StreamDevSwitch() operation will be skipped.
Due to this mDevices will remain as speaker and when SSR is up,
active streams will route to the speaker only.
Thus in this corner scenario, update the mDevices as BT A2DP so that
when SSR is up active streams will be routed to BT properly.

Change-Id: I35c6a6c095a83f32dfba17b939dabcdeff67f0ef