pal: handling for SCO when DEVICE_NONE routing comes while dev is disconnected
If a stream is active on SCO and playback has ended, APM will send routing=0.
Stream will be closed in PAL after standby time. If SCO device gets disconnected,
this stream will not receive new routing and stream will remain with SCO for the
time being. If SCO device gets connected again with different config in the
meantime and capture stream tries to start ABR path, it will lead to error due to
config mismatch.

Added handling to route SCO streams to Speaker during ReleaseAudiPatch when SCO
device is disconnected.

Change-Id: I59d0b785054c5761b97ecd768877c09010385391