PAL: Manage UPD backend device during force/stream device switch
While UPD is running on speaker in concurrency with music playback,
when a wired-headset/Bluetooth device is connected, the UPD stream
skips the switch to these devices; but, continues to run on speaker
under NLPI. This change, ensures that the UPD is switched to handset
(LPI) under this usecase.
In addition, when UPD is running in concurrency with voice call on
handset device, on inserting wired-headset or connecting BT device,
the UPD skips switching to these devices and continues to run on
handset device. This change ensures that the handset power mode is
set accordingly under these circumstances.

Change-Id: I438e63a41dcdf0e2f04d452b082e36eedfd12ba1