pal: default initialize isAbrEnabled flag to false
isAbrEnabled flag is not reset in case of any failure before the start
of ABR path. If after the failure codec switch happens and the codec
does not support ABR, it will incorrectly try to find device kv pair
with ABR enabled. This is because we update device metadata once before
we get the encoder configurations.

To avoid this scenario, set isAbrEnabled flag to false at the beiginning
of first BT playback session start.

Change-Id: If8cedb9a4c53c44a92e932a78838c87b78297b91