pal: move sco out device to the beginning of device disconnect list
We move all active streams to SCO when BT_SCO=on is received. Streams on
SCO devices are also disconnected and connected again to handle config
changes of SCO device. There might be scenarios where few streams are
already active on SCO during BT_SCO=on handling, and if we disconnect
other devices first and try to connect the streams to SCO, it might lead
to error in cases where earlier SCO was started with different config.

Fixed this by moving SCO device to the beginning of the disconnect list.
This will ensure that SCO devices with old configs will be disconnected
first before getting connected with the new configs.

Change-Id: Iefc9f0a7045315a1d9c747d240236575f6c934d2