Bluetooth: query latency from bt_offload on every request
For APTx codecs, BT latency changes while the usecase is going
on based on transport and the usecase. Latency values needs to
be queried on every call from framework. Query latency values
from BT IPC on every call to report it back to framework. Remove
dependency on plugins to update the latency values.

Change-Id: Ibb97231b0dd99823e4d7065df85e91f0b7782ba9