pal: Start Recognition during Device switch for Active Stream
In case of dual VA, if the first stream is in buffering state,
and second stream is in loaded state.
It will get connected when connect event is called for last
stream but as the last stream will remain in loaded state after
device switch is handled, detection for first stream will never
get started as engine will not be in Active State.

Now, ensuring that if any stream was in active state before switch,
we will start engine, while handling connection for last stream.

Change-Id: I47582f1794991a7c1c11f041b3268b43e812faac