PAL: Close graph in failure case while getting instance ID for tag
Graph was opened for to get wakeup module's version, but in case
of failure in getting instance ID for module tag, graph was not
closed, due to which later if same GKV was opened, no new graph
was getting opened, due to which during close also, nothing was
closed, and while unmapping we get failure, as graph was still
there.

Close graph in case of failure in getting instance ID for module tag.

Change-Id: I51828e1f04642f66726f708630ca9f5ef6c7921e